+++ FloconAndroid/okhttp-interceptor/src/main/java/io/github/openflocon/flocon/okhttp/OkHttpInterceptor.kt
@@ -9,11 +9,7 @@ import okhttp3.Interceptor
 import okhttp3.MediaType
 import okhttp3.Request
 import okhttp3.Response
-import okio.Buffer
-import okio.GzipSource
 import java.io.IOException
-import java.nio.charset.Charset
-import java.nio.charset.StandardCharsets
 import java.util.UUID
 
 data class FloconNetworkIsImageParams(
@@ -49,12 +45,12 @@ class FloconOkhttpInterceptor(
             request.headers.toMultimap().mapValues { it.value.joinToString(",") }
 
         if (requestBody != null) {
-            val (rBodyString, rSize) = extractRequestBodyInfo(
+            val (bodyString, bodySize) = extractRequestBodyInfo(
                 request = request,
                 requestHeaders = requestHeadersMap,
             )
-            requestBodyString = rBodyString
-            requestSize = rSize
+            requestBodyString = bodyString
+            requestSize = bodySize
         }
 
         val startTime = System.nanoTime()
@@ -114,12 +110,12 @@ class FloconOkhttpInterceptor(
                 response.headers.toMultimap().mapValues { it.value.joinToString(",") }
 
             if (responseBody != null) {
-                val (rBodyString, rSize) = extractResponseBodyInfo(
+                val (bodyString, bodySize) = extractResponseBodyInfo(
                     response = response,
                     responseHeaders = responseHeadersMap,
                 )
-                responseBodyString = rBodyString
-                responseSize = rSize
+                responseBodyString = bodyString
+                responseSize = bodySize
             }
 
             val isImage =
@@ -191,66 +187,4 @@ class FloconOkhttpInterceptor(
         }
     }
 
-}
-
-internal fun MediaType?.charsetOrUtf8(): Charset = this?.charset() ?: Charsets.UTF_8
-
-private fun extractResponseBodyInfo(
-    response: Response,
-    responseHeaders: Map<String, String>
-): Pair<String?, Long?> {
-    val responseBody = response.body ?: return null to null
-
-    var bodyString: String? = null
-    var bodySize: Long? = null
-
-    val source = responseBody.source()
-    source.request(Long.MAX_VALUE) // Buffer the entire body, otherwise we have an empty string
-
-    var buffer = source.buffer
-
-    val charset = responseBody.contentType().charsetOrUtf8()
-
-    val contentEncoding = responseHeaders["Content-Encoding"] ?: responseHeaders["content-encoding"]
-    if ("gzip".equals(contentEncoding, ignoreCase = true)) {
-        GzipSource(buffer.clone()).use { gzippedResponseBody ->
-            buffer = Buffer()
-            buffer.writeAll(gzippedResponseBody)
-        }
-
-        bodyString = buffer.clone().readString(charset)
-        bodySize = buffer.size
-    } else {
-        bodyString = buffer.clone().readString(charset)
-        bodySize = buffer.size
-    }
-
-    return bodyString to bodySize
-}
-
-private fun extractRequestBodyInfo(
-    request: Request,
-    requestHeaders: Map<String, String>
-): Pair<String?, Long?> {
-    val requestBody = request.body ?: return null to null
-
-    var bodySize: Long? = null
-
-    var buffer = Buffer()
-    requestBody.writeTo(buffer)
-
-    val contentEncoding = requestHeaders["Content-Encoding"] ?: requestHeaders["content-encoding"]
-
-    if ("gzip".equals(contentEncoding, ignoreCase = true)) {
-        bodySize = buffer.size
-        GzipSource(buffer).use { gzippedResponseBody ->
-            buffer = Buffer()
-            buffer.writeAll(gzippedResponseBody)
-        }
-    }
-
-    val charset = requestBody.contentType().charsetOrUtf8()
-    val bodyString = buffer.readString(charset)
-
-    return bodyString to bodySize
 }
